refactor: use import syntax
