+++ test/scripts/types/enum.js
@@ -1,11 +1,9 @@
-'use strict';
-
-require('chai').should();
-const ValidationError = require('../../../built/error/validation');
+import chai from 'chai';
+const should = chai.should(); // eslint-disable-line
+import ValidationError from '../../../lib/error/validation';
+import SchemaTypeEnum from '../../../lib/types/enum';
 
 describe('SchemaTypeEnum', () => {
-  const SchemaTypeEnum = require('../../../built/types/enum');
-
   it('validate()', () => {
     const type = new SchemaTypeEnum('test', {elements: ['foo', 'bar', 'baz']});
