+++ test/scripts/types/number.js
@@ -1,10 +1,9 @@
-'use strict';
-
-require('chai').should();
-const ValidationError = require('../../../built/error/validation');
+import chai from 'chai';
+const should = chai.should(); // eslint-disable-line
+import ValidationError from '../../../lib/error/validation';
+import SchemaTypeNumber from '../../../lib/types/number';
 
 describe('SchemaTypeNumber', () => {
-  const SchemaTypeNumber = require('../../../built/types/number');
   const type = new SchemaTypeNumber('type');
 
   it('cast()', () => {
